Add initial rewrite
