+++ pubspec.yaml
@@ -57,6 +57,8 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  mobx: ^2.0.4
+  flutter_mobx: ^2.0.2
 
 dev_dependencies:
   flutter_test:
@@ -64,6 +66,7 @@ dev_dependencies:
   flutter_launcher_icons: ^0.9.0
   json_serializable: ^4.1.0
   build_runner: ^1.11.1
+  mobx_codegen: ^2.0.2
 
 flutter_icons:
   android: true
